devops/requirements.txt
chevron
boto3
pyyaml
fabric
invoke
filetype
